requirements.txt
streamlit==1.42.0
numpy==2.2.2
opencv-python-headless==4.11.0.86
Pillow==11.1.0
reportlab==4.3.1
